Invert swap probability check
As the probability approaches 1 we want more swaps to occur, creating a
more random graph. As it approaches 1 less numbers in the range 0-1 are
greater than it, so the previous (>) check reduced the number of swaps.
This also means that giving a probability of 0 gets you a regular graph
as expected, and matching the Watts-Strogatz paper.