Allow number of edges per vertex to be specified
I don't know if this is actually required for us, but it's a parameter
in the Watts-Strogatz paper (k). I imagine it could be useful for some
comparisons with (eg.) social networks where k would model the number of
friends the average person has. Also cleans up some memory management
stuff & checks that the command line arguments are valid.